Making some modifications

+++ src/main/java/com/advjava/hospitalmanagement/controllers/AuthController.java
@@ -68,7 +68,7 @@ public class AuthController {
     @GetMapping("/me")
     public ResponseEntity<UserDto> me() {
         var authentication= SecurityContextHolder.getContext().getAuthentication();
-        var userId= (Long)authentication.getPrincipal();
+        var userId= (Integer) authentication.getPrincipal();
         var user = userRepository.findById(userId).orElse(null);
         if (user == null) {
             return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
